--- data/impl/C_Timer.After.lua
@@ -1,2 +1,0 @@
-local time, seconds, callback = ...
-time.newTicker(seconds, callback, 1)
