--- data/impl/C_Timer.NewTicker.lua
@@ -1,2 +1,0 @@
-local time, seconds, callback, iterations = ...
-return time.newTicker(seconds, callback, iterations or 1)
