--- data/impl/C_Timer.NewTimer.lua
@@ -1,2 +1,0 @@
-local time, seconds, callback = ...
-return time.newTicker(seconds, callback, 1)
