--- data/impl/GetTime.lua
@@ -1,2 +1,0 @@
-local time = ...
-return time.stamp
