+++ data/impl/RequestTimePlayed.lua
@@ -1,4 +1,4 @@
-local api, time = ...
-time.timers:push(time.stamp, function()
+local api = ...
+api.modules.time.AddTimer(0, function()
   api.SendEvent('TIME_PLAYED_MSG', 3600, 600)
 end)
