+++ data/impl/SendSystemMessage.lua
@@ -1,4 +1,4 @@
-local api, time, units, msg = ...
-time.timers:push(time.stamp, function()
+local api, units, msg = ...
+api.modules.time.AddTimer(0, function()
   api.SendEvent('CHAT_MSG_SYSTEM', msg, units.guids[units.aliases.player].name, nil, '')
 end)
